Crate stdmath

Modules

core

exclude

Macros

ctx

stage_default_methods

Structs

Product

Sigma

Enums

Method
Method by which to process combinatorics

Traits

One
Defines a multiplicative identity element for Self.

Pow
Trait for pow-supported numbers.

Zero
Defines a additive identity element for Self.

Functions

binomial
Returns the binomial expansion of the equation (a + b) ^ n

combination
Returns the computed combination nCr

factorial
Returns the computed factorial of a number

factorial_count
Returns the number of digits of a factorial computation

pascals
Returns a vector of vectors of representing n layers of a binomial triangle

permutation
Returns the computed permutation nPr

product
Returns the product of functionally transformed items from a range

ramanujansPI
Returns the value of PI using Ramanujan’s Formula

sigma
Returns the summation of functionally transformed items from a range