Trait stdmath::core::Resolve

pub trait Resolve: Simplify {
    type Result;
    fn _resolve(self: Box<Self>) -> Self::Result;

    fn as_any(&self) -> &dyn Any;

    fn is_friendly_with(
        &self, 
        other: &dyn Resolve<Result = Self::Result>
    ) -> bool;

    fn _cmp(
        &self, 
        other: &dyn Resolve<Result = Self::Result>
    ) -> Option<Ordering>;

    fn _hash(&self, _state: &mut dyn Hasher);

    fn _to_context(self: Box<Self>) -> Context<Self::Result>;

    fn _as_context(self: Box<Self>) -> Context<Self::Result>;

    fn resolve(self) -> Self::Result
    where
        Self: Sized,
    { ... }

    fn _clone(&self) -> Box<dyn Resolve<Result = Self::Result>> { ... }

    fn _debug(&self, _f: &mut Formatter<'_>) -> Result { ... }

    fn to_context(self) -> Context<Self::Result>
    where
        Self: Sized,
    { ... }

    fn as_context(self) -> Context<Self::Result>
    where
        Self: Sized,
    { ... }
}

Trait for objects that can be resolved to a value.

How can I implement Resolve?

Resolve primarily requires the implementation of the resolve method. But also depends on all methods being implemented. The default behaviour of a provided method is to panic from unimplementation (See unimplemented!).

You can use the helpful stage_default_methods! macro to apply default implementations of these methods.

Resolve requries that your type implements Simplify.

use std::fmt;
use stdmath::{stage_default_methods, core::{Simplify, Resolve}};

#[derive(Debug, Hash, Clone, PartialEq, PartialOrd)]
struct Add(u8, u8);

impl Simplify for Add {
    fn simplify(&self, file: &mut dyn fmt::Write) -> fmt::Result {
        write!(file, "({} + {})", self.0, self.1)
    }
}

impl Resolve for Add {
    type Result = u8;
    stage_default_methods!(is_friendly_with to_context ALL);
    fn resolve(self: Box<Self>) -> Self::Result {
        self.0 + self.1
    }
}

let val = Add(10, 20);
assert_eq!(val.repr().expect("failed to represent context"), "(10 + 20)");
assert_eq!(Box::new(val).resolve(), 30);

Associated Types

type Result

The result of the object resolution.

Loading content...
Required methods

fn _resolve(self: Box<Self>) -> Self::Result

This method resolves self into it's equivalent Result type.

fn as_any(&self) -> &dyn Any

This method transposes self into a dynamic Any representation.

struct MyValue;
impl Resolve for MyValue {
    ...
    fn as_any(&self) -> &dyn Any {
        // cast a reference to self as a dynamic reference
        self
    }
    ...
}

This is required for dynamicism between varied types that impl Resolve.

fn is_friendly_with(&self, other: &dyn Resolve<Result = Self::Result>) -> bool

This method determines whether the type behind the dynamic reference other is a type that is friendly with self during simplification & resolution.

What this means

During simplification, friendly values can be included within the same bracket scope.

Given (1 + 2 + (3 + 4))

2 and 1 are friendly

3 and 2 are unfriendly

4 and 3 are friendly

Example

struct MyValue;
impl Resolve for MyValue {
    ...
    fn is_friendly_with(&self, other: &dyn Resolve<Result = Self::Result>) -> bool {
        // check if other is a valid object whose type matches MyValue
        other.as_any().is::<Self>()
    }
    ...
}

This is required for dynamicism between varied types that impl Resolve.

fn _cmp(&self, other: &dyn Resolve<Result = Self::Result>) -> Option<Ordering>

This method partially compares self with the type behind the dynamic reference other

Partial Comparison

returns None if other is not a type that matches Self.

returns None if other is not partially comparable to Self.

returns Some(Ordering) otherwise.

Example

use std::cmp::PartialOrd;

#[derive(PartialEq, PartialOrd)]
struct MyValue;
impl Resolve for MyValue {
    ...
    fn _cmp(&self, other: &dyn Resolve<Result = Self::Result>) -> Option<Ordering> {
        // attempt casting other to the type MyValue
        // after conversion, partially compare both types
        other
            .as_any()
            .downcast_ref::<Self>()
            .map_or(None, |other| PartialOrd::partial_cmp(self, other))
    }
    ...
}

This is required for dynamicism between varied types that impl Resolve.

fn _hash(&self, _state: &mut dyn Hasher)

This method enables hashing trait objects that implements Resolve

Example

use std::hash::{Hash, Hasher};

#[derive(Hash)]
struct MyValue;
impl Resolve for MyValue {
    ...
    fn _hash(&self, state: &mut dyn Hasher) {
        // proxy for hashing Self
        Hash::hash(self, &mut state)
    }
    ...
}

This is required for dynamicism between varied types that impl Resolve.

fn _to_context(self: Box<Self>) -> Context<Self::Result>

This method allows wrapping a resolvable value into a context

Behaviour

If the underlying value is already a context, it returns that, otherwise, wrap it within a Nil context

Example

struct MyValue;
impl Resolve for MyValue {
    ...
    fn to_context(self) -> Context<Self::Result> {
        // wrap self within a context
        stdmath::core::value(self)
    }
    ...
}

This is required for dynamicism between varied types that impl Resolve.

fn _as_context(self: Box<Self>) -> Context<Self::Result>

Loading content...
Provided methods

fn resolve(self) -> Self::Result where
Self: Sized,

fn _clone(&self) -> Box<dyn Resolve<Result = Self::Result>>

This method enables cloning trait objects that implements Resolve

Example

#[derive(Clone)]
struct MyValue;
impl Resolve for MyValue {
    ...
    fn _clone(&self) -> Box<dyn Resolve<Result = Self::Result>> {
        // return a Resolve trait object from a clone of Self
        Box::new(self.clone()) as Box<dyn Resolve<Result = Self::Result>>
    }
    ...
}

While this is a provided method, it's default implementation is to panic from unimplementation (See unimplemented!). So, if cloning functionality is ever needed, you'd have to overload this method.

fn _debug(&self, _f: &mut Formatter<'_>) -> Result

This method enables debugging trait objects that implements Resolve

Example

use std::fmt;

#[derive(Debug)]
struct MyValue;
impl Resolve for MyValue {
    ...
    fn _debug(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
        // proxy for debugging Self
        fmt::Debug::fmt(self, f)
    }
    ...
}

While this is a provided method, it's default implementation is to panic from unimplementation (See unimplemented!). So, if debugging is ever needed, you'd have to overload this method.

fn to_context(self) -> Context<Self::Result> where
Self: Sized,

fn as_context(self) -> Context<Self::Result> where
Self: Sized,

Loading content...